+++ package.json
@@ -26,6 +26,8 @@
     "file-loader": "1.1.5",
     "fs-extra": "3.0.1",
     "gh-pages": "^1.2.0",
+    "glamor": "^2.20.40",
+    "glamorous": "^4.13.1",
     "html-webpack-plugin": "2.29.0",
     "jest": "20.0.4",
     "mock-socket": "^8.0.0",
@@ -43,6 +45,7 @@
     "react-virtualized": "^9.20.0",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
+    "reselect": "^3.0.1",
     "resolve": "1.6.0",
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
