+++ src/app.js
@@ -2,11 +2,8 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-import Layout from './components/layout'
-import ListWrap from './components/list-wrap'
-import Mocks from './containers/mocks'
-import Chats from './containers/chats'
-import { initSocket, closeSocket } from './redux/modules/socket'
+import {ChatsPage} from "./chats/pages/chats";
+import { initSocket, closeSocket } from './ducks/socket/socket'
 
 class App extends Component {
   static propTypes = {
@@ -24,12 +21,7 @@ class App extends Component {
 
   render() {
     return (
-      <Layout>
-        <Mocks />
-        <ListWrap>
-          <Chats />
-        </ListWrap>
-      </Layout>
+     <ChatsPage/>
     );
   }
 };
