+++ src/chats/components/avatar/avatar.scss
