+++ src/chats/components/time.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { getDateObj, isToday } from '../../utils'
 
 export default class Time extends PureComponent {
-  getFormattedTime(timestamp) {
+  getFormattedTime = (timestamp) => {
     const messageDate = new Date(timestamp);
     const dateObj = getDateObj(messageDate);
     const timeStr = `${dateObj.hours}:${dateObj.minutes}`;
