+++ src/ducks/chats/types.js
@@ -1,0 +1,2 @@
+export const CHATS_UPDATE_ACTION = 'CHATS_LOADED_ACTION';
+export const CHAT_UPDATE_ACTION = 'CHAT_LOADED_ACTION';
