+++ src/ducks/user/actions.js
@@ -1,0 +1,8 @@
+import { CURRENT_USER_UPDATE } from "./types";
+
+export const userUpdateAction = (payload) => {
+  return {
+    type: CURRENT_USER_UPDATE,
+    payload
+  };
+};
