+++ src/ducks/user/index.js
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './actions';
+export * from './operations';
+export * from './reducer';
