+++ src/ducks/user/operations.js
@@ -1,0 +1,8 @@
+import { userUpdateAction } from "./actions";
+import api from "../../api/chats";
+
+
+export const loadUser = () => async (dispatch) => {
+  const user = await api.getUser();
+  return dispatch(userUpdateAction(user));
+};
