+++ src/ducks/user/types.js
@@ -1,0 +1,1 @@
+export const CURRENT_USER_UPDATE = 'CURRENT_USER_UPDATE';
